Crate re_ui

Rerun GUI theme and helpers, built around egui.

Re-exports

pub use self::icons::Icon;

Modules

command 🔒

command_palette 🔒

context_ext 🔒

design_tokens 🔒

drag_and_drop

Helpers for drag and drop support for reordering hierarchical lists.

icons

list_item

Second-generation list item.

markdown_utils 🔒

modal

section_collapsing_header 🔒

syntax_highlighting 🔒

toasts

A toast notification system for egui, roughly based on https://github.com/urholaukkarinen/egui-toast.

ui_ext 🔒

Structs

CommandPalette

DesignTokens

The look and feel of the UI.

KeyMarkdown

Newtype over [egui::Key] which provides a Display implementation suitable for markdown.

ModifiersMarkdown

Newtype over [egui::Modifiers] which provides a Display implementation suitable for markdown.

MouseButtonMarkdown

Newtype over [egui::PointerButton] which provides a Display implementation suitable for markdown.

SectionCollapsingHeader

A collapsible section header, with support for optional help tooltip and button.

TopBarStyle

Enums

LabelStyle

The style of a label.

UICommand

All the commands we support.

Constants

CUSTOM_WINDOW_DECORATIONS

If true, we hide the native window decoration (the top bar with app title, close button etc), and instead paint our own close/maximize/minimize buttons.

FULLSIZE_CONTENT

If true, we fill the entire window, except for the close/maximize/minimize buttons in the top-left. See https://github.com/emilk/egui/pull/2049

NATIVE_WINDOW_BAR

If true, we show the native window decorations/chrome with the close/maximize/minimize buttons and app title.

Traits

ContextExt

Extension trait for [egui::Context].

SyntaxHighlighting

UICommandSender

Interface for sending UICommand messages.

UiExt

Rerun custom extensions to [egui::Ui].

Functions

apply_style_and_install_loaders

Apply the Rerun design tokens to the given egui context and install image loaders.

design_tokens

Return a reference to the global design tokens structure.

format_with_decimals_in_range 🔒

is_in_resizable_panel 🔒

Is this Ui in a resizable panel?