Module re_ui::drag_and_drop

Helpers for drag and drop support for reordering hierarchical lists.

Works well in combination with crate::list_item::ListItem.

Structs

DropTarget

Drop target information, including where to draw the drop indicator and where to insert the dragged item.

ItemContext

Context information about the hovered item.

Enums

ItemKind

Functions

find_drop_target

Compute the geometry of the drag cursor and where the dragged item should be inserted.