Module re_ui::full_span

Support for full-span widgets.

Full-span widgets are widgets which draw beyond the boundaries of ui.max_rect(), e.g. to provide highlighting without margin, like crate::list_item2. Typically, in the context of a side panel, the full span is the entire width of the panel, excluding any margin.

This module maintains a stack of full span values (effectively[egui::Rangef]) using nestable scopes (via full_span_scope) and makes them available to widgets (via get_full_span).

Structs

FullSpanStack 🔒

Functions

full_span_scope

Set up a full-span scope.

get_full_span

Retrieve the current full-span scope.