re_ui::icon_text
Function shortcut_with_icon

pub fn shortcut_with_icon(
    ctx: &Context,
    modifiers: Modifiers,
    icon: impl Into<IconText>,
) -> IconText

Shows a keyboard shortcut with a modifier and the given icon.

On Mac, this will show the symbol for the modifier. Otherwise, it will show the name of the modifier, and a “+” between the modifier and the icon.