Module re_ui::list_item

Second-generation list item.

TODO(ab): provide some top-level documentation here.

Modules

button_content 🔒

item_button 🔒

Abstraction for buttons to be used in list items.

label_content 🔒

list_item 🔒

Core list item functionality.

other_contents 🔒

property_content 🔒

scope 🔒

Structs

ButtonContent

Simple ListItemContent to easily display a button in a crate::list_item::ListItem-based UI.

ContentContext

Context provided to ListItemContent implementations

CustomContent

ListItemContent that delegates to a closure.

DebugContent

ListItemContent that displays the content rect.

ItemActionButton

An super::ItemButton that acts as an action button.

ItemMenuButton

An super::ItemButton that acts as a popup menu.

LabelContent

ListItemContent that displays a simple label with optional icon and buttons.

LayoutInfo

Layout information prepared by list_item_scope to be used by super::ListItemContent.

ListItem

Content-generic list item.

PropertyContent

ListItemContent to display property-like, two-column content, with the left column containing a label (along with an optional icon) and the right column containing some custom value (which may be editable).

ShowCollapsingResponse

Responses returned by ListItem::show_hierarchical_with_children.

Enums

DesiredWidth

Traits

ItemButton

A button that can be used by a list item.

ListItemContent

Functions

list_item_scope

Create a scope in which [ListItem]s can be created.