Struct re_ui::list_item::ListItem

pub struct ListItem<'a> {
    text: WidgetText,
    re_ui: &'a ReUi,
    active: bool,
    selected: bool,
    subdued: bool,
    force_hovered: bool,
    collapse_openness: Option<f32>,
    width_allocation_mode: WidthAllocationMode,
    icon_fn: Option<Box<dyn FnOnce(&ReUi, &mut Ui, Rect, WidgetVisuals) + 'a>>,
    buttons_fn: Option<Box<dyn FnOnce(&ReUi, &mut Ui) -> Response + 'a>>,
}

Generic widget for use in lists.

Layout:

┌───┬────────────────────────────────────────────────────────────┬───┐
│   │┌──────┐ ┌──────┐                           ┌──────┐┌──────┐│   │
│   ││  __  │ │      │                           │      ││      ││   │
│   ││  \/  │ │ icon │  label                    │ btns ││ btns ││   │
│   ││      │ │      │                           │      ││      ││   │
│   │└──────┘ └──────┘                           └──────┘└──────┘│   │
└───┴────────────────────────────────────────────────────────────┴───┘
    ◀───────────── allocated width (used for layout) ───────────▶
◀────────────── clip rectangle (used for highlighting) ─────────────▶

Features:

selectable

full span highlighting based on clip rectangle

optional icon

optional on-hover buttons on the right

optional collapsing behavior for trees

This widget relies on the clip rectangle to be properly set as it use it for the shape if its background highlighting. This has a significant impact on the hierarchy of the UI. This is typically how things should be laid out:

Panel (no margin, set the clip rectangle)
└── ScrollArea (no margin)
    └── Frame (with inner margin)
        └── ListItem

See ReUi::panel_content for an helper to build the [egui::Frame] with proper margins.

Fields
text: WidgetTextre_ui: &'a ReUiactive: boolselected: boolsubdued: boolforce_hovered: boolcollapse_openness: Option<f32>width_allocation_mode: WidthAllocationModeicon_fn: Option<Box<dyn FnOnce(&ReUi, &mut Ui, Rect, WidgetVisuals) + 'a>>buttons_fn: Option<Box<dyn FnOnce(&ReUi, &mut Ui) -> Response + 'a>>
Implementations

impl<'a> ListItem<'a>

pub fn new(re_ui: &'a ReUi, text: impl Into<WidgetText>) -> Self

Create a new ListItem with the given label.

pub fn active(self, active: bool) -> Self

Set the active state the item.

pub fn selected(self, selected: bool) -> Self

Set the selected state of the item.

pub fn subdued(self, subdued: bool) -> Self

Set the subdued state of the item.

pub fn force_hovered(self, force_hovered: bool) -> Self

Override the hovered state even if the item is not actually hovered.

Used to highlight items representing things that are hovered elsewhere in the UI. Note that the [egui::Response] returned by Self::show and ]Self::show_collapsing] will still reflect the actual hover state.

pub fn width_allocation_mode(self, mode: WidthAllocationMode) -> Self

Set the width allocation mode.

pub fn with_icon(self, icon: &'a Icon) -> Self

Provide an Icon to be displayed on the left of the item.

pub fn with_icon_fn( self, icon_fn: impl FnOnce(&ReUi, &mut Ui, Rect, WidgetVisuals) + 'a ) -> Self

Provide a custom closure to draw an icon on the left of the item.

pub fn with_buttons( self, buttons: impl FnOnce(&ReUi, &mut Ui) -> Response + 'a ) -> Self

Provide a closure to display on-hover buttons on the right of the item.

Notes:

If buttons are used, the item will allocate the full available width of the parent. If the enclosing UI adapts to the childrens width, it will unnecessarily grow. If buttons aren’t used, the item will only allocate the width needed for the text and icons if any.

A right to left layout is used, so the right-most button must be added first.

pub fn show(self, ui: &mut Ui) -> Response

Draw the item.

pub fn show_collapsing<R>( self, ui: &mut Ui, id: Id, default_open: bool, add_body: impl FnOnce(&ReUi, &mut Ui) -> R ) -> ShowCollapsingResponse<R>

Draw the item as a collapsing header.

fn ui(self, ui: &mut Ui, id: Option<Id>) -> ListItemResponse