Module re_ui::list_item2

Second-generation list item.

TODO(ab): provide some top-level documentation here.

Modules

label_content 🔒

list_item 🔒

Core list item functionality.

other_contents 🔒

property_content 🔒

scope 🔒

Structs

ContentContext

Context provided to ListItemContent implementations

CustomContent

ListItemContent that delegates to a closure.

DebugContent

ListItemContent that displays the content rect.

LabelContent

ListItemContent that displays a simple label with optional icon and buttons.

LayoutInfo

Layout information prepared by list_item_scope to be used by super::ListItemContent.

ListItem

Content-generic list item.

PropertyContent

ListItemContent to display property-like, two-column content, with the left column containing a label (along with an optional icon) and the right column containing some custom value (which may be editable).

ShowCollapsingResponse

Responses returned by ListItem::show_hierarchical_with_children.

Enums

DesiredWidth

Traits

ListItemContent

Functions

list_item_scope

Create a scope in which [ListItem]s can be created.