re_ui
Module time_drag_value

Structs

TimeDragValue

Drag value widget for editing time values for both sequence and temporal timelines.

Statics

SPAN_TO_START_TIME_OFFSET_THRESHOLD 🔒

Value of the start time over time span ratio above which an explicit offset is handled.

Functions

round_down 🔒

round_up 🔒

time_range_base_time 🔒

unit_from_span 🔒