Crate re_ui

Rerun GUI theme and helpers, built around egui.

Re-exports

pub use icons::Icon;

Modules

command 🔒

command_palette 🔒

design_tokens 🔒

egui_helpers

icons

layout_job_builder 🔒

list_item

toasts

A toast notification system for egui, roughly based on https://github.com/urholaukkarinen/egui-toast.

toggle_switch 🔒

Adapted from egui_demo_lib/src/demo/toggle_switch.rs

Structs

CommandPalette

DesignTokens

The look and feel of the UI.

LayoutJobBuilder

Utility for building layout jobs.

ReUi

TopBarStyle

Enums

UICommand

All the commands we support.

Constants

CUSTOM_WINDOW_DECORATIONS

If true, we hide the native window decoration (the top bar with app title, close button etc), and instead paint our own close/maximize/minimize buttons.

FULLSIZE_CONTENT

If true, we fill the entire window, except for the close/maximize/minimize buttons in the top-left. See https://github.com/emilk/egui/pull/2049

NATIVE_WINDOW_BAR

If true, we show the native window decorations/chrome with the close/maximize/minimize buttons and app title.

Traits

UICommandSender

Interface for sending UICommand messages.

Functions

help_hover_button

native_window_buttons_ui

Show some close/maximize/minimize buttons for the native window.

toggle_switch

iOS-style toggle switch.