re_ui
Module filter_widget

Structs

FilterMatcher

Full-text, case-insensitive matcher.

FilterState

State and UI for the filter widget.

InnerState 🔒

State for the filter widget when it is toggled on.

Keyword 🔒

A single keyword from a query.

MatchStateMachine 🔒

State machine used to test a given keyword against a given sequence of path parts.

MergeRanges 🔒

Given a vector of ranges, iterate over the sorted, merged ranges.

PathRanges

Accumulates highlight ranges for the various parts of a path.

Enums

MatchState 🔒

Functions

format_matching_text

Given a list of highlight sections defined by start/end indices and a string, produce a properly highlighted [egui::WidgetText].

single_keyword_matches 🔒

Helper function to extract all matches of a given keyword in the given text.