Struct re_ui::ReUi

pub struct ReUi {
    pub egui_ctx: Context,
    pub design_tokens: DesignTokens,
    pub static_image_cache: Arc<Mutex<StaticImageCache>>,
}

Fields
egui_ctx: Contextdesign_tokens: DesignTokens

Colors, styles etc loaded from a design_tokens.json

static_image_cache: Arc<Mutex<StaticImageCache>>
Implementations

impl ReUi

pub fn load_and_apply(egui_ctx: &Context) -> Self

Create ReUi and apply style to the given egui context.

pub fn rerun_logo(&self) -> Arc<RetainedImage>

pub fn view_padding() -> f32

Margin on all sides of views.

pub fn window_rounding() -> f32

pub fn normal_rounding() -> f32

pub fn small_rounding() -> f32

pub fn table_line_height() -> f32

pub fn table_header_height() -> f32

pub fn top_bar_margin() -> Margin

pub fn top_bar_height() -> f32

Height of the top-most bar.

pub fn title_bar_height() -> f32

Height of the title row in the blueprint view and selection view, as well as the tab bar height in the viewport view.

pub fn native_window_rounding() -> f32

pub fn top_panel_frame(&self) -> Frame

pub fn bottom_panel_margin(&self) -> Vec2

pub fn bottom_panel_frame(&self) -> Frame

For the streams view (time panel)

pub fn small_icon_size() -> Vec2

pub fn setup_table_header(_header: &mut TableRow<'_, '_>)

pub fn setup_table_body(body: &mut TableBody<'_>)

pub fn warning_text(&self, text: impl Into<String>) -> RichText

pub fn error_text(&self, text: impl Into<String>) -> RichText

pub fn loop_selection_color() -> Color32

The color we use to mean “loop this selection”

pub fn loop_everything_color() -> Color32

The color we use to mean “loop all the data”

pub fn paint_watermark(&self)

Paint a watermark

pub fn top_bar_style(
&self,
native_pixels_per_point: Option<f32>,
fullscreen: bool
) -> TopBarStyle

pub fn icon_image(&self, icon: &Icon) -> Arc<RetainedImage>

pub fn small_icon_button(&self, ui: &mut Ui, icon: &Icon) -> Response

pub fn medium_icon_toggle_button(
&self,
ui: &mut Ui,
icon: &Icon,
selected: &mut bool
) -> Response

pub(crate) fn large_button_impl(
&self,
ui: &mut Ui,
icon: &Icon,
bg_fill: Option<Color32>,
tint: Option<Color32>
) -> Response

pub fn large_button(&self, ui: &mut Ui, icon: &Icon) -> Response

pub fn large_button_selected(
&self,
ui: &mut Ui,
icon: &Icon,
selected: bool
) -> Response

pub fn visibility_toggle_button(
&self,
ui: &mut Ui,
visible: &mut bool
) -> Response

pub fn large_collapsing_header<R>(
&self,
ui: &mut Ui,
label: &str,
default_open: bool,
add_body: impl FnOnce(&mut Ui) -> R
)

pub fn grid_left_hand_label(&self, ui: &mut Ui, label: &str) -> Response

Workaround for putting a label into a grid at the top left of its row.

pub fn selection_grid(&self, ui: &mut Ui, id: &str) -> Grid

Two-column grid to be used in selection view.

pub fn draw_shadow_line(&self, ui: &mut Ui, rect: Rect, direction: Direction)

Draws a shadow into the given rect with the shadow direction given from dark to light

pub fn selectable_label_with_icon(
&self,
ui: &mut Ui,
icon: &Icon,
text: impl Into<WidgetText>,
selected: bool
) -> Response