Trait re_ui::ContextExt

pub trait ContextExt {
    // Required method
    fn ctx(&self) -> &Context;

    // Provided methods
    fn text_format_body(&self) -> TextFormat { ... }
    fn text_format_key(&self) -> TextFormat { ... }
    fn rerun_logo_uri(&self) -> &'static str { ... }
    fn hover_stroke(&self) -> Stroke { ... }
    fn selection_stroke(&self) -> Stroke { ... }
    fn warning_text(&self, text: impl Into<String>) -> RichText { ... }
    fn error_text(&self, text: impl Into<String>) -> RichText { ... }
    fn top_bar_style(&self, style_like_web: bool) -> TopBarStyle { ... }
    fn paint_watermark(&self) { ... }
}

Extension trait for [egui::Context].

This trait provides Rerun-specific helpers and utilities that require access to the egui context.

Required Methods

fn ctx(&self) -> &Context

Provided Methods

fn text_format_body(&self) -> TextFormat

Text format used for regular body.

fn text_format_key(&self) -> TextFormat

Text format used for labels referring to keys and buttons.

fn rerun_logo_uri(&self) -> &'static str

fn hover_stroke(&self) -> Stroke

Hovered UI and spatial primitives should have this outline.

fn selection_stroke(&self) -> Stroke

Selected UI and spatial primitives should have this outline.

fn warning_text(&self, text: impl Into<String>) -> RichText

fn error_text(&self, text: impl Into<String>) -> RichText

fn top_bar_style(&self, style_like_web: bool) -> TopBarStyle

fn paint_watermark(&self)

Paint a watermark

Object Safety

This trait is not object safe.